+++ PolynomesAlgo/PolynomesAlgo/Monome.h
@@ -7,12 +7,12 @@ class Monome
 {
 	//valeur, exposant, suivant
 public:
-	Monome(float value, float exposant);
+	Monome(float value, int exposant);
 	Monome* PrimitiveMonome();
 	Monome* DeriveMonome();
 
 	float value;
-	float exposant;
+	int exposant;
 	Monome* next;
 };
